# avoid a one extra object encryption key copy in tryDecryptedETag