## Description
avoid a double copy in tryDecryptedETag

## Motivation and Context
ListObjects return value is reliant on tryDecryptedETag to
be faster, this PR retries to reduce an extra allocation
by re-using the object encryption key already obtained as
part of decryptObjectInfo()

## How to test this PR?
Nothing special all encryption functionalities should work

## Types of changes
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
- [ ] Fixes a regression (If yes, please add `commit-id` or `PR #` here)
- [ ] Documentation needed
- [ ] Unit tests needed