@harshavardhana 2020-10-19
Okay so found out that main tiny allocations happen in `cmd/crypto/key.UnsealETag`  @klauspost it is harder to change that, I found a way to gain some performance though by using bufferpool

Not much though 

```
~ benchcmp etag new-etag 
benchmark                       old ns/op     new ns/op     delta
BenchmarkTryDecryptedETag-4     26559         26043         -1.94%

benchmark                       old allocs     new allocs     delta
BenchmarkTryDecryptedETag-4     32             29             -9.38%

benchmark                       old bytes     new bytes     delta
BenchmarkTryDecryptedETag-4     143408        141317        -1.46%
```